% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{check_version_prospect}
\alias{check_version_prospect}
\title{This function checks if the input parameters are defined as expected
to run either PROSPECT-D or PROSPECT-PRO}
\usage{
check_version_prospect(LMA, PROT, CBC)
}
\arguments{
\item{LMA}{numeric. content corresponding to LMA}

\item{PROT}{numeric. content corresponding to protein content}

\item{CBC}{numeric. content corresponding to carbon based constituents}
}
\value{
list. updated LMA, PROT and CBC
}
\description{
This function checks if the input parameters are defined as expected
to run either PROSPECT-D or PROSPECT-PRO
}
