% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_band.R
\name{closest_band}
\alias{closest_band}
\title{This function identifies the closest spectral band among a selection of bands}
\usage{
closest_band(target_wl, wl)
}
\arguments{
\item{target_wl}{numeric. central wavelength of spectral band of interest}

\item{wl}{numeric. vector of spectral bands}
}
\value{
selected_band rank of selected band
}
\description{
This function identifies the closest spectral band among a selection of bands
}
