% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_to_ansi.R
\name{colour_to_ansi}
\alias{colour_to_ansi}
\title{Convert plain text colour to ANSI code}
\usage{
colour_to_ansi(colour)
}
\arguments{
\item{colour}{colour in plain text ("red", "green", etc.) to convert to ANSI}
}
\value{
string representing provided colour as ANSI encoding
}
\description{
Convert plain text colour to ANSI code
}
\examples{
colour_to_ansi("red") # gives: "\033[31m"
}
