% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input_prospect.R
\name{complete_input_prospect}
\alias{complete_input_prospect}
\title{Complete the list of PROSPECT parameters with default values}
\usage{
complete_input_prospect(input_prospect, parm_to_add, expected_parms)
}
\arguments{
\item{input_prospect}{input parameters sent to PROSPECT by user}

\item{parm_to_add}{Parameters to be added to input parameters}

\item{expected_parms}{full set of parameters expected to run PROSPECT}
}
\value{
input_prospect
}
\description{
Complete the list of PROSPECT parameters with default values
}
