% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{CostVal_RMSE_ABS}
\alias{CostVal_RMSE_ABS}
\title{Value of the cost criterion to minimize during PROSPECT inversion}
\usage{
CostVal_RMSE_ABS(RT, Abs)
}
\arguments{
\item{RT}{list. Simulated reflectance and transmittance}

\item{Abs}{numeric. Absorptance on which PROSPECT ins inverted}
}
\value{
fc sum of squared difference between simulated and measured leaf optical properties
}
\description{
Value of the cost criterion to minimize during PROSPECT inversion
}
