% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_input_prospect.R
\name{define_input_prospect}
\alias{define_input_prospect}
\title{This function produces a data frame from all prospect input variables if not
defined already}
\usage{
define_input_prospect(
  input_prospect,
  chl = NULL,
  car = NULL,
  ant = NULL,
  brown = NULL,
  ewt = NULL,
  lma = NULL,
  prot = NULL,
  cbc = NULL,
  n_struct = NULL,
  alpha = NULL
)
}
\arguments{
\item{input_prospect}{numeric. prospect input parameters}

\item{chl}{numeric. Chlorophyll content (microg.cm-2)}

\item{car}{numeric. carotenoid content (microg.cm-2)}

\item{ant}{numeric. anthocyanin content (microg.cm-2)}

\item{brown}{numeric. brown pigment content (Arbitrary units)}

\item{ewt}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{lma}{numeric. Leaf Mass per Area (g.cm-2)}

\item{prot}{numeric. protein content  (g.cm-2)}

\item{cbc}{numeric. Nonprot carbon-based constituent content (g.cm-2)}

\item{n_struct}{numeric. Leaf structure parameter}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf}
}
\value{
list. updated lma, prot and cbc
}
\description{
This function produces a data frame from all prospect input variables if not
defined already
}
