% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_leaf_db.R
\name{download_leaf_db}
\alias{download_leaf_db}
\title{Complete the list of PROSPECT parameters with default values}
\usage{
download_leaf_db(url_db = NULL, db_name = "ANGERS")
}
\arguments{
\item{url_db}{character. URL for online repository where to download data}

\item{db_name}{character. name of the database available online}
}
\value{
list. Includes leaf chemistry, refl, tran & number of samples
}
\description{
Complete the list of PROSPECT parameters with default values
}
