% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{download_LeafDB}
\alias{download_LeafDB}
\title{Complete the list of PROSPECT parameters with default values}
\usage{
download_LeafDB(urldb = NULL, dbName = "ANGERS")
}
\arguments{
\item{urldb}{character. URL for online repository where to download data}

\item{dbName}{character. name of the database available online}
}
\value{
list. Includes leaf chemistry, refl, tran & number of samples
}
\description{
Complete the list of PROSPECT parameters with default values
}
