\name{EAZY_filters}
\alias{EAZY_filters}
\docType{data}
\title{
EAZY Filter Set
}
\description{
A re-organised version of the EAZY filter set as downloaded on 20/11/2018. The idea is that subsets of this can be directly passed into the various \code{\link{SFH}} functions.
}
\usage{data("EAZY_filters")}
\format{
  The format is:\cr
List of 2\cr
 $ info   : chr [1:347] "hst/ACS_update_sep07/wfc_f435w_t77.dat" "hst/ACS_update_sep07/wfc_f475w_t77.dat" "hst/ACS_update_sep07/wfc_f555w_t77.dat" "hst/ACS_update_sep07/wfc_f606w_t77.dat" ...\cr
 $ filters:List of 347\cr
  ..$ : num [1:133, 1:2] 3565 3575 3585 3595 3605 ...\cr
  ..$ : num [1:121, 1:2] 3825 3840 3855 3870 3885 ...\cr
  ..$ : num [1:118, 1:2] 4515 4530 4545 4560 4575 ...\cr
  ..$ : num [1:173, 1:2] NA 4600 4615 4630 4645 ...\cr
  ..$ : num [1:86, 1:2] NA 6535 6560 6585 6610 ...\cr
  ..$ : num [1:117, 1:2] NA 6770 6795 6820 6845 ...\cr
  ..$ : num [1:102, 1:2] NA 7860 7890 7920 7950 7980 8010 8040 8070 8100 ...\cr
  ..$ : num [1:193, 1:2] NA 7740 7775 7810 7845 ...\cr
  ..$ : num [1:110, 1:2] NA 13460 13510 13560 13610 ...\cr
  ..$ : num [1:1409, 1:2] NA 2310 2315 2320 2325 ...\cr
  ..$ : num [1:469, 1:2] NA 2920 2930 2940 2950 2960 2970 2980 2990 3000 ...\cr
  ..$ : num [1:183, 1:2] 3545 3560 3575 3590 3605 ...\cr
  ..$ : num [1:208, 1:2] 4355 4370 4385 4400 4415 ...\cr
  ..$ : num [1:142, 1:2] NA 4640 4660 4680 4700 4720 4740 4760 4780 4800 ...\cr
  ..$ : num [1:149, 1:2] NA 5815 5835 5855 5875 ...\cr
  ..$ : num [1:125, 1:2] NA 6860 6885 6910 6935 ...\cr
  ..$ : num [1:102, 1:2] NA 7940 7970 8000 8030 8060 8090 8120 8150 8180 ...\cr
  ..$ : num [1:94, 1:2] NA 29900 30015 30130 30245 ...\cr
  ..$ : num [1:86, 1:2] NA 38050 38200 38350 38500 ...\cr
  ..$ : num [1:95, 1:2] NA 48630 48820 49010 49200 ...\cr
  ..$ : num [1:158, 1:2] NA 60200 60460 60720 60980 ...\cr
  ..$ : num [1:178, 1:2] 3000 3020 3040 3060 3080 3100 3120 3140 3160 3180 ...\cr
  ..$ : num [1:237, 1:2] 2980 3000 3020 3040 3060 3080 3100 3120 3140 3160 ...\cr
  ..$ : num [1:86, 1:2] NA 5660 5680 5700 5720 5740 5760 5780 5800 5820 ...\cr
  ..$ : num [1:116, 1:2] NA 6890 6915 6940 6965 ...\cr
  ..$ : num [1:38, 1:2] NA 10600 10700 10800 10900 11000 11100 11200 11300 11400 ...\cr
  ..$ : num [1:43, 1:2] NA 14400 14500 14600 14700 14800 14900 15000 15100 15200 ...\cr
  ..$ : num [1:62, 1:2] NA 19000 19100 19200 19300 19400 19500 19600 19700 19800 ...\cr
  ..$ : num [1:50, 1:2] NA 19300 19400 19500 19600 19700 19800 19900 20000 20100 ...\cr
  ..$ : num [1:112, 1:2] NA 9925 9965 10005 10045 ...\cr
  ..$ : num [1:67, 1:2] NA 10960 11000 11040 11080 ...\cr
  ..$ : num [1:79, 1:2] NA 14475 14525 14575 14625 ...\cr
  ..$ : num [1:67, 1:2] NA 19265 19335 19405 19475 ...\cr
  ..$ : num [1:111, 1:2] NA 9945 9985 10025 10065 ...\cr
  ..$ : num [1:94, 1:2] NA 9970 10010 10050 10090 ...\cr
  ..$ : num [1:78, 1:2] NA 14500 14550 14600 14650 ...\cr
  ..$ : num [1:66, 1:2] NA 19310 19380 19450 19520 ...\cr
  ..$ : num [1:900, 1:2] NA 2990 3000 3010 3020 3030 3040 3050 3060 3070 ...\cr
  ..$ : num [1:761, 1:2] NA 3025 3035 3045 3055 ...\cr
  ..$ : num [1:173, 1:2] 3015 3025 3035 3045 3055 ...\cr
  ..$ : num [1:141, 1:2] 3285 3295 3305 3315 3325 ...\cr
  ..$ : num [1:526, 1:2] NA 3010 3020 3030 3040 3050 3060 3070 3080 3090 ...\cr
  ..$ : num [1:141, 1:2] 3285 3295 3305 3315 3325 ...\cr
  ..$ : num [1:227, 1:2] 3530 3540 3550 3560 3570 3580 3590 3600 3610 3620 ...\cr
  ..$ : num [1:236, 1:2] 3535 3545 3555 3565 3575 ...\cr
  ..$ : num [1:193, 1:2] 3915 3930 3945 3960 3975 ...\cr
  ..$ : num [1:227, 1:2] 3530 3540 3550 3560 3570 3580 3590 3600 3610 3620 ...\cr
  ..$ : num [1:144, 1:2] 4595 4610 4625 4640 4655 ...\cr
  ..$ : num [1:139, 1:2] 4635 4650 4665 4680 4695 ...\cr
  ..$ : num [1:81, 1:2] 4795 4810 4825 4840 4855 ...\cr
  ..$ : num [1:144, 1:2] 4595 4610 4625 4640 4655 ...\cr
  ..$ : num [1:210, 1:2] 5495 5515 5535 5555 5575 ...\cr
  ..$ : num [1:210, 1:2] 5495 5515 5535 5555 5575 ...\cr
  ..$ : num [1:142, 1:2] NA 5525 5545 5565 5585 ...\cr
  ..$ : num [1:210, 1:2] 5495 5515 5535 5555 5575 ...\cr
  ..$ : num [1:92, 1:2] 6985 7010 7035 7060 7085 ...\cr
  ..$ : num [1:93, 1:2] 6985 7010 7035 7060 7085 ...\cr
  ..$ : num [1:148, 1:2] NA 7285 7310 7335 7360 ...\cr
  ..$ : num [1:98, 1:2] 7020 7045 7070 7095 7120 ...\cr
  ..$ : num [1:98, 1:2] 7795 7825 7855 7885 7915 ...\cr
  ..$ : num [1:98, 1:2] 7795 7825 7855 7885 7915 ...\cr
  ..$ : num [1:98, 1:2] 7795 7825 7855 7885 7915 ...\cr
  ..$ : num [1:98, 1:2] 7795 7825 7855 7885 7915 ...\cr
  ..$ : num [1:46, 1:2] NA 11400 11450 11500 11550 ...\cr
  ..$ : num [1:74, 1:2] NA 14450 14500 14550 14600 ...\cr
  ..$ : num [1:272, 1:2] NA 13980 14050 14120 14190 ...\cr
  ..$ : num [1:85, 1:2] NA 10390 10430 10470 10510 ...\cr
  ..$ : num [1:80, 1:2] NA 14570 14620 14670 14720 ...\cr
  ..$ : num [1:62, 1:2] NA 19455 19525 19595 19665 ...\cr
  ..$ : num [1:76, 1:2] 4295 4310 4325 4340 4355 ...\cr
  ..$ : num [1:204, 1:2] 4230 4255 4280 4305 4330 ...\cr
  ..$ : num [1:269, 1:2] 2960 2980 3000 3020 3040 3060 3080 3100 3120 3140 ...\cr
  ..$ : num [1:47, 1:2] 2980 3005 3030 3055 3080 ...\cr
  ..$ : num [1:89, 1:2] 3630 3655 3680 3705 3730 ...\cr
  ..$ : num [1:75, 1:2] 5380 5405 5430 5455 5480 ...\cr
  ..$ : num [1:100, 1:2] 6565 6585 6605 6625 6645 ...\cr
  ..$ : num [1:134, 1:2] 7805 7830 7855 7880 7905 ...\cr
  ..$ : num [1:161, 1:2] 3630 3640 3650 3660 3670 3680 3690 3700 3710 3720 ...\cr
  ..$ : num [1:107, 1:2] 4810 4825 4840 4855 4870 ...\cr
  ..$ : num [1:118, 1:2] 3850 3865 3880 3895 3910 ...\cr
  ..$ : num [1:92, 1:2] 5400 5420 5440 5460 5480 5500 5520 5540 5560 5580 ...\cr
  ..$ : num [1:94, 1:2] 6720 6745 6770 6795 6820 ...\cr
  ..$ : num [1:81, 1:2] 7925 7955 7985 8015 8045 ...\cr
  ..$ : num [1:14, 1:2] 7950 7975 8000 8025 8050 ...\cr
  ..$ : num [1:142, 1:2] NA 10240 10280 10320 10360 ...\cr
  ..$ : num [1:74, 1:2] NA 14565 14615 14665 14715 ...\cr
  ..$ : num [1:110, 1:2] NA 16985 17055 17125 17195 ...\cr
  ..$ : num [1:132, 1:2] NA 2960 2980 3000 3020 3040 3060 3080 3100 3120 ...\cr
  ..$ : num [1:241, 1:2] 2960 2980 3000 3020 3040 3060 3080 3100 3120 3140 ...\cr
  ..$ : num [1:230, 1:2] NA 3640 3660 3680 3700 3720 3740 3760 3780 3800 ...\cr
  ..$ : num [1:139, 1:2] NA 5450 5475 5500 5525 ...\cr
  ..$ : num [1:152, 1:2] NA 6470 6495 6520 6545 ...\cr
  ..$ : num [1:126, 1:2] NA 3210 3220 3230 3240 3250 3260 3270 3280 3290 ...\cr
  ..$ : num [1:219, 1:2] 3415 3425 3435 3445 3455 ...\cr
  ..$ : num [1:153, 1:2] NA 4680 4695 4710 4725 ...\cr
  ..$ : num [1:199, 1:2] NA 5495 5515 5535 5555 ...\cr
  ..$ : num [1:87, 1:2] NA 6870 6895 6920 6945 ...\cr
  ..$ : num [1:178, 1:2] NA 3200 3210 3220 3230 3240 3250 3260 3270 3280 ...\cr
  ..$ : num [1:76, 1:2] 4510 4525 4540 4555 4570 ...\cr
  .. [list output truncated]\cr
}
\details{
The info component of the list contains the main path of the original filter used, from which it is possible to work out the filter type and telescope etc. This is kept close to the form supplied in EAZY, and in the same order.

The filters list contains the corresponding wavelength (in Angstroms) and the relative throughout response.
}
\source{
https://github.com/gbrammer/eazy-photoz/tree/master/filters
}
\references{
Brammer, van Dokkum & Coppi, 2008, ApJ, 686, 1503
}
\examples{
data(EAZY_filters)
plot(EAZY_filters$filters[[158]], type='l', xlab = 'Wave / Ang', ylab = 'Response')

#Make filter subset which can be passed to \option{filters} argument in \code{\link{SFH}}:

SDSS_filters=EAZY_filters$filters[grep('sdss',EAZY_filters$info)]
plot(SDSS_filters[[1]], type='l', xlim=c(3e3,1.1e4), ylim=c(0,40), xlab = 'Wave / Ang',
ylab = 'Response')
for(i in 2:5){
  lines(SDSS_filters[[i]])
}

#Generate a SFH using some EAZY filters:

EAZY_filters$info[c(120:121,278:282,256:259)]

SFHdemo_dust=SFHfunc(filters=EAZY_filters$filters[c(120:121,278:282,256:259)])
plot(SFHdemo_dust$out[,2:3], log='x', ylim=c(22,17), xlab=BC03lr$Labels$Wavelab,
ylab='Mag')
}
\keyword{datasets}
