% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spectral_data.R
\name{fit_spectral_data}
\alias{fit_spectral_data}
\title{This function adapts spec_prospect accordingly to experimental data
or to a spectral domain defined by user_domain}
\usage{
fit_spectral_data(
  lambda,
  spec_prospect = NULL,
  refl = NULL,
  tran = NULL,
  user_domain = NULL,
  ul_bounds = FALSE
)
}
\arguments{
\item{lambda}{numeric. Spectral bands corresponding to experimental data}

\item{spec_prospect}{list. Includes optical constants: refractive index,
specific absorption coefficients and corresponding spectral bands}

\item{refl}{numeric. Measured reflectance data}

\item{tran}{numeric. Measured transmittance data}

\item{user_domain}{numeric. either Lower and upper bounds for domain of
interest (optional) or list of spectral bands of interest}

\item{ul_bounds}{boolean. set to TRUE if user_domain only includes lower and
upper band, set to FALSE if user_domain is a list of spectral bands (in nm)}
}
\value{
list including spectral properties at the new resolution
}
\description{
This function adapts spec_prospect accordingly to experimental data
or to a spectral domain defined by user_domain
}
