% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_prior.R
\name{get_N_prior}
\alias{get_N_prior}
\title{This function defines a regression model to estimate N from R only or T only}
\usage{
get_N_prior(
  lambda,
  spec_prospect = NULL,
  refl = NULL,
  tran = NULL,
  OptWL_R = list(NIR = 800, SWIR = 1131),
  OptWL_T = list(NIR = 753, SWIR = 1121)
)
}
\arguments{
\item{lambda}{numeric. spectral bands corresponding to experimental data}

\item{spec_prospect}{list. Includes optical constants refractive index,
specific absorption coefficients and corresponding spectral bands}

\item{refl}{numeric. Measured reflectance data}

\item{tran}{numeric. Measured Transmittance data}

\item{OptWL_R}{list. optimal wavelengths used to estimate N from R only}

\item{OptWL_T}{list. optimal wavelengths used to estimate N from T only}
}
\value{
Nprior prior estimation of N with R or T only
}
\description{
This function defines a regression model to estimate N from R only or T only
}
