% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_PROSPECT-data.R
\docType{data}
\name{input_prospect}
\alias{input_prospect}
\title{Outcomes of 10 combinations of input parameters for PROSPECT}
\format{
A data frame with 10 rows and 6 variables:
\itemize{
\item \emph{chl}: Chlorophyll content
\item \emph{car}: carotenoid content
\item \emph{ant}: anthocyanin content
\item \emph{ewt}: Equivalent water thickness
\item \emph{lma}: Leaf Mass per Area
\item \emph{n_struct}: leaf structure parameter
}
}
\usage{
input_prospect
}
\description{
corresponds to input parameters from PROSPECT-D:
}
\keyword{datasets}
