% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_prospect.R
\name{invert_prospect}
\alias{invert_prospect}
\title{Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
invert_prospect(
  refl = NULL,
  tran = NULL,
  parms_to_estimate = "ALL",
  prospect_version = "D",
  options = NULL
)
}
\arguments{
\item{refl}{numeric. Measured reflectance data}

\item{tran}{numeric. Measured Transmittance data}

\item{parms_to_estimate}{character. Parameters to estimate (can be 'ALL')}

\item{prospect_version}{character. prospect version used for inversion: 'D' or 'PRO'
See details.}

\item{options}{list.
\itemize{
\item init_values data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.
\item merit_function character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)
\item xlub data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{parms_to_estimate}
first line being the lower boundaries and second line the upper boundaries.
\item estimate_brown_pigments boolean. should brown pigments be accounted for during inversion?
\item estimate_alpha boolean. should alpha be accounted for during inversion?
\item verbose boolean. set true to get info about adjustment of tolerance or initialization
\item progressBar boolean. show progressbar?
}}
}
\value{
output_prospect estimated values corresponding to parms_to_estimate
}
\description{
Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
\details{
Two versions of prospect are available for inversion.
The version is depending on the parameters taken into account:\tabular{ccc}{
   Version \tab D \tab PRO \cr
   chl \tab ✅ \tab ✅ \cr
   car \tab ✅ \tab ✅ \cr
   ant \tab ✅ \tab ✅ \cr
   brown \tab ✅ \tab ✅ \cr
   ewt \tab ✅ \tab ✅ \cr
   lma \tab ✅ \tab  \cr
   prot \tab  \tab ✅ \cr
   cbc \tab  \tab ✅ \cr
   n_struct \tab ✅ \tab ✅ \cr
}


Argument \code{init_values} is expecting a default value for each of the parameters as well as an \code{alpha} value.
}
