% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_prospect_opt.R
\name{invert_prospect_opt}
\alias{invert_prospect_opt}
\title{This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only reflectance or only Transmittance is available}
\usage{
invert_prospect_opt(
  lambda,
  refl = NULL,
  tran = NULL,
  parms_to_estimate = "ALL",
  prospect_version = "D",
  options = NULL
)
}
\arguments{
\item{lambda}{numeric. spectral bands corresponding to experimental data
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{refl}{numeric. Measured reflectance data}

\item{tran}{numeric. Measured transmittance data}

\item{parms_to_estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{prospect_version}{character. Version of prospect model used for the
inversion: 'D', 'PRO'.}

\item{options}{list.
\itemize{
\item init_values data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.
\item merit_function character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)
\item xlub data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{parms_to_estimate}
first line being the lower boundaries and second line the upper boundaries.
\item estimate_brown_pigments boolean. should brown pigments be accounted for during inversion?
\item estimate_alpha boolean. should alpha be accounted for during inversion?
\item verbose boolean. set true to get info about adjustment of tolerance or initialization
\item progressBar boolean. show progressbar?
}}
}
\value{
Nprior vector corresponding to teh prior estimation of N based on R only or T only
}
\description{
This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only reflectance or only Transmittance is available
}
