% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_prospect_subdomain.R
\name{invert_prospect_subdomain}
\alias{invert_prospect_subdomain}
\title{Function performing PROSPECT inversion on a spectral subset combining an
initial feature subset (initial_features) and an additional feature subset (additional_features)}
\usage{
invert_prospect_subdomain(
  additional_features,
  refl,
  tran,
  lambda,
  measured_bioch,
  target_parm,
  parms_to_estimate,
  initial_features,
  prospect_version = "D",
  options = NULL
)
}
\arguments{
\item{additional_features}{spectral bands (in nm) to be added to the initial_features}

\item{refl}{numeric. matrix of reflectances (n spectral bands x p samples)}

\item{tran}{numeric. matrix of transmittances (n spectral bands x p samples)}

\item{lambda}{numeric. spectral bands corresponding to reflectance and transmittance measurements}

\item{measured_bioch}{numeric. value of biophysical/biochemical parameter to estimate for each sample}

\item{target_parm}{character. name of the parameter.
Should be picked between "chl", "car", "ant", "brown", "ewt", "prot", "cbc", "n_struct"}

\item{parms_to_estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{initial_features}{Initial feature set (in nm)}

\item{prospect_version}{character. Version of prospect model used for the inversion: 'D' or 'PRO'}

\item{options}{list.
\itemize{
\item init_values data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.
\item merit_function character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)
\item xlub data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate}
first line being the lower boundaries and second line the upper boundaries.
\item estimate_brown_pigments boolean. should brown pigments be accounted for during inversion?
\item estimate_alpha boolean. should alpha be accounted for during inversion?
\item verbose boolean. set true to get info about adjustment of tolerance or initialization
\item progressBar boolean. show progressbar?
}}
}
\value{
list containing estimated value of the target and RMSE compared to measured value
}
\description{
Function performing PROSPECT inversion on a spectral subset combining an
initial feature subset (initial_features) and an additional feature subset (additional_features)
}
