% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merit_prospect_drmse.R
\name{merit_prospect_drmse}
\alias{merit_prospect_drmse}
\title{Merit function for PROSPECT inversion}
\usage{
merit_prospect_drmse(
  x,
  spec_prospect,
  refl,
  tran,
  input_prospect,
  parms_to_estimate
)
}
\arguments{
\item{x}{numeric. Vector of input variables to estimate}

\item{spec_prospect}{list. Includes optical constants refractive index,
specific absorption coefficients and corresponding spectral bands}

\item{refl}{numeric. measured reflectance data}

\item{tran}{numeric. measured Transmittance data}

\item{input_prospect}{dataframe. set of PROSPECT input variables}

\item{parms_to_estimate}{numeric. location of variables from input_prospect
to be estimated through inversion}
}
\value{
fc estimates of the parameters
}
\description{
Merit function for PROSPECT inversion
}
