% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Merit_RMSE_PROSPECT_ABS}
\alias{Merit_RMSE_PROSPECT_ABS}
\title{Merit function for PROSPECT inversionm using absorptance instead of Reflectance and Transmittance}
\usage{
Merit_RMSE_PROSPECT_ABS(xinit, SpecPROSPECT, Abs, Input_PROSPECT,
  WhichVars2Estimate)
}
\arguments{
\item{xinit}{numeric. Vector of input variables to estimate}

\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Abs}{numeric. measured Transmittance data}

\item{Input_PROSPECT}{dataframe. set of PROSPECT input variables}

\item{WhichVars2Estimate}{numeric. location of variables from Input_PROSPECT
to be estimated through inversion}
}
\value{
fc estimates of the parameters
}
\description{
Merit function for PROSPECT inversionm using absorptance instead of Reflectance and Transmittance
}
