% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptDomain_RT-data.R
\docType{data}
\name{OptDomain_RT}
\alias{OptDomain_RT}
\title{OptDomain_RT: optimal spectral domains defined to assess the different leaf
chemical constituents.
These optimal spectral domains were defined in
Féret et al. (2019) https://doi.org/10.1016/j.rse.2018.11.002
Féret et al. (2021) https://doi.org/10.1016/j.rse.2020.112173
Spafford et al. (2021) https://doi.org/10.1016/j.rse.2020.112176}
\format{
A list including sequence of spectral bands corresponding to each
leaf constituent
\itemize{
\item \emph{CHL}: optimal spectral domain for CHL
\item \emph{CAR}: optimal spectral domain for CAR
\item \emph{ANT}: optimal spectral domain for ANT
\item \emph{EWT}: optimal spectral domain for EWT
\item \emph{LMA}: optimal spectral domain for LMA
\item \emph{PROT}: optimal spectral domain for PROT
\item \emph{CBC}: optimal spectral domain for CBC
}
}
\usage{
OptDomain_RT
}
\description{
OptDomain_RT: optimal spectral domains defined to assess the different leaf
chemical constituents.
These optimal spectral domains were defined in
Féret et al. (2019) https://doi.org/10.1016/j.rse.2018.11.002
Féret et al. (2021) https://doi.org/10.1016/j.rse.2020.112173
Spafford et al. (2021) https://doi.org/10.1016/j.rse.2020.112176
}
\keyword{datasets}
