% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_domains_rt-data.R
\docType{data}
\name{optimal_domains_rt}
\alias{optimal_domains_rt}
\title{optimal_domains_rt: optimal spectral domains defined to assess the different leaf
chemical constituents.
These optimal spectral domains were defined in
Féret et al. (2019) https://doi.org/10.1016/j.rse.2018.11.002
Féret et al. (2021) https://doi.org/10.1016/j.rse.2020.112173
Spafford et al. (2021) https://doi.org/10.1016/j.rse.2020.112176}
\format{
A list including sequence of spectral bands corresponding to each
leaf constituent
\itemize{
\item \emph{chl}: optimal spectral domain for chl
\item \emph{car}: optimal spectral domain for car
\item \emph{ant}: optimal spectral domain for ant
\item \emph{ewt}: optimal spectral domain for ewt
\item \emph{lma}: optimal spectral domain for lma
\item \emph{prot}: optimal spectral domain for prot
\item \emph{cbc}: optimal spectral domain for cbc
}
}
\usage{
optimal_domains_rt
}
\description{
optimal_domains_rt: optimal spectral domains defined to assess the different leaf
chemical constituents.
These optimal spectral domains were defined in
Féret et al. (2019) https://doi.org/10.1016/j.rse.2018.11.002
Féret et al. (2021) https://doi.org/10.1016/j.rse.2020.112173
Spafford et al. (2021) https://doi.org/10.1016/j.rse.2020.112176
}
\keyword{datasets}
