% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_features_sfs.R
\name{optimal_features_sfs}
\alias{optimal_features_sfs}
\title{Function performing optimal feature selection based on sequential forward feature slection}
\usage{
optimal_features_sfs(
  refl = NULL,
  tran = NULL,
  lambda,
  measured_bioch,
  target_parm,
  parms_to_estimate = "ALL",
  spectral_domain,
  spectral_width,
  number_features,
  prospect_version = "D",
  nbCPU = 1,
  continue = FALSE,
  already_done = NULL,
  options = NULL
)
}
\arguments{
\item{refl}{numeric. matrix of reflectances (n spectral bands x p samples)}

\item{tran}{numeric. matrix of transmittances (n spectral bands x p samples)}

\item{lambda}{numeric. spectral bands corresponding to reflectance and transmittance measurements}

\item{measured_bioch}{numeric. value of biophysical/biochemical parameter to estimate for each sample}

\item{target_parm}{character. name of the parameter.
Should be picked between "chl", "car", "ant", "brown", "ewt", "prot", "cbc", "n_struct"}

\item{parms_to_estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{spectral_domain}{vector. defines minimum and maximum wavelengths of the spectral domain (in nm).
Assumes 1nm spectral sampling}

\item{spectral_width}{vector. width of individual spectral features (in nm)}

\item{number_features}{vector. number of features to be identified}

\item{prospect_version}{character.}

\item{nbCPU}{numeric. defines number of CPU for multithread processing}

\item{continue}{boolean. set to TRUE if the function has already been run for a lower value of number_features}

\item{already_done}{list. contains output of function optimal_features_SFS previouly run with lower value of number_features}

\item{options}{list.}
}
\value{
list containing :
\itemize{
\item spectral_features = the Spectral Features identified as most relevant for estimation of target_parm (vector)
\item spectral_features_List = the Spectral Features identified as most relevant for estimation of target_parm (list)
\item estimated_parms = the estimated target_parm parameter for each additional spectral feature
\item RMSE = the minimum RMSE corresponding to the estimation of target_parm for each asditional spectral feature
}
}
\description{
Function performing optimal feature selection based on sequential forward feature slection
}
