% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{plotinv}
\alias{plotinv}
\title{Function plotting results of PROSPECT inversion for a given biophysical property}
\usage{
plotinv(
  BP_df,
  Labs = NULL,
  MinMax = NULL,
  Colors = NULL,
  stats = TRUE,
  filename = NULL
)
}
\arguments{
\item{BP_df}{dataframe. should include fields 'estimated', 'measured', and 'config'}

\item{Labs}{character. labels for X and Y axes}

\item{MinMax}{numeric. min and max axis values}

\item{Colors}{character. colors corresponding to the different configurations in BP_df}

\item{stats}{boolean. should statistics be displayed on figure}

\item{filename}{character. path for the file to be saved}
}
\value{
plotxy ggplot object
}
\description{
Function plotting results of PROSPECT inversion for a given biophysical property
}
