% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{print_msg}
\alias{print_msg}
\title{Function printing message
\itemize{
\item if leaf optics do not match with spectral domain
\item if leaf optics do not match with expected data class
\item if confusion is identified for prospect version to use
\item if missing input variables during PROSPECT_LUT simulations
}}
\usage{
print_msg(cause, args = NULL)
}
\arguments{
\item{cause}{character. cause raising message print}

\item{args}{list. optional arguments used to print message}
}
\value{
none
}
\description{
Function printing message
\itemize{
\item if leaf optics do not match with spectral domain
\item if leaf optics do not match with expected data class
\item if confusion is identified for prospect version to use
\item if missing input variables during PROSPECT_LUT simulations
}
}
