% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{print_msg_classLOP}
\alias{print_msg_classLOP}
\title{Function priniting message if leaf optics do not match with expected data class}
\usage{
print_msg_classLOP()
}
\value{
none
}
\description{
Function priniting message if leaf optics do not match with expected data class
}
