% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{print_msg_lengthLOP}
\alias{print_msg_lengthLOP}
\title{Function priniting message if leaf optics do not match with spectral domain}
\usage{
print_msg_lengthLOP()
}
\value{
none
}
\description{
Function priniting message if leaf optics do not match with spectral domain
}
