\name{ProFiltTrans}
\alias{ProFiltTrans}
\docType{data}
\title{
Pre-computed Magnitudes for Random SFHs
}
\description{
A list of pre-computed magnitudes over a range of lookback time (separated in Gyr intervals) for randomly sampled SFHs using \code{\link{massfunc_b5}} and Uniform sampling (from 0 to 1 Msol/yr) for each mass bin. These outputs can be used to compute filter conversions without the added cost and complexity of creating \code{\link{ProSpectSED}} SEDs.
}
\usage{data("ProFiltTrans")}
\format{
  The format is:
List of 6
 $ Names  : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
 $ maglist:List of 13
  ..$ Age0 : num [1:1000, 1:341] 19.8 18.2 20.2 17.7 18.6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age1 : num [1:1000, 1:341] 21.8 22.4 21.9 21.8 21.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age2 : num [1:1000, 1:341] 22.9 23.9 23 23.3 23.1 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age3 : num [1:1000, 1:341] 23.8 24.9 23.9 24.3 24.1 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age4 : num [1:1000, 1:341] 24.6 25.8 24.7 25.2 24.9 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age5 : num [1:1000, 1:341] 25.4 26.5 25.4 26 25.7 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age6 : num [1:1000, 1:341] 26.1 27.3 26 26.8 26.5 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age7 : num [1:1000, 1:341] 26.7 27.8 26.6 27.4 27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age8 : num [1:1000, 1:341] 27.7 28.5 27.7 28.2 27.9 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age9 : num [1:1000, 1:341] 28.7 29.4 28.9 29.3 28.8 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age10: num [1:1000, 1:341] 33.8 31.6 33.8 31.6 32.3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age11: num [1:1000, 1:341] 36.2 36 36.5 35.9 35.9 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
  ..$ Age12: num [1:1000, 1:341] 40.8 39.5 41.7 39.1 39.6 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:1000] "result.1" "result.2" "result.3" "result.4" ...
  .. .. ..$ : chr [1:341] "RestUV/Tophat_2700_200.dat" "hst/wfc3/UVIS/f275w.dat" "REST_FRAME/UV2800.dat" "RestUV/Tophat_2800_200.dat" ...
 $ pivwave: num [1:341] 2699 2708 2791 2800 2992 ...
 $ params : num [1:1000, 1:6] 0.774 0.197 0.978 0.201 0.361 ...
 $ zsteps : num [1:14] 0 0.0737 0.1561 0.249 0.3553 ...
 $ ages   : int [1:14] 0 1 2 3 4 5 6 7 8 9 ...
}
\details{
All available filters in \code{ProSpect} that cover u band to the K band are included (341 filters). This means almost all useful filter conversions can be computed using this, a sensible age window and \code{\link{filterTranMags}} (see Examples).
}
\examples{
filterTranMags(ProFiltTrans$maglist$Age3[,c("g_SDSS", "r_SDSS", "i_SDSS")],
               ProFiltTrans$maglist$Age3[,"r_VST"])
}
\keyword{datasets}
