% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect-package.R
\docType{package}
\name{prospect-package}
\alias{prospect}
\alias{prospect-package}
\title{prospect: This package includes PROSPECT leaf model as well as inversion routines for the estimation of leaf chemistry from spectroscopic measurements}
\description{
The package prospect includes PROSPECT-PRO, the latest version of the PROSPECT leaf model. It also includes 
              inversion routines based on iterative optimization in order to estimate leaf chemistry from spectroscopic measurements.
}
\author{
\strong{Maintainer}: Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (0000-0002-0151-1334) [copyright holder]

Authors:
\itemize{
  \item Florian de Boissieu \email{florian.deboissieu@irstea.fr} (0000-0002-2185-9952)
}

}
\keyword{internal}
