% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{PROSPECT}
\alias{PROSPECT}
\title{core function running PROSPECT
This function allows simulations using PROSPECT-D or PROSPECT-PRO depending
on the parameterization.}
\usage{
PROSPECT(
  SpecPROSPECT = NULL,
  Input_PROSPECT = NULL,
  N = 1.5,
  CHL = 40,
  CAR = 8,
  ANT = 0,
  BROWN = 0,
  EWT = 0.01,
  LMA = NULL,
  PROT = 0,
  CBC = 0,
  alpha = 40,
  check = TRUE
)
}
\arguments{
\item{SpecPROSPECT}{list. Includes spectral constants derived from
SpecPROSPECT_FullRange: refractive index, specific absorption coefficients
and corresponding spectral bands}

\item{Input_PROSPECT}{list. Includes all prospect input parameters}

\item{N}{numeric. Leaf structure parameter}

\item{CHL}{numeric. Chlorophyll content (microg.cm-2)}

\item{CAR}{numeric. Carotenoid content (microg.cm-2)}

\item{ANT}{numeric. Anthocyanin content (microg.cm-2)}

\item{BROWN}{numeric. Brown pigment content (Arbitrary units)}

\item{EWT}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{LMA}{numeric. Leaf Mass per Area (g.cm-2)}

\item{PROT}{numeric. protein content  (g.cm-2)}

\item{CBC}{numeric. NonProt Carbon-based constituent content (g.cm-2)}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf}

\item{check}{boolean. set to TRUE to check input data format}
}
\value{
leaf directional-hemispherical reflectance and transmittance
}
\description{
core function running PROSPECT
This function allows simulations using PROSPECT-D or PROSPECT-PRO depending
on the parameterization.
}
