% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_lut.R
\name{prospect_lut}
\alias{prospect_lut}
\title{computation of a LUT of leaf optical properties using a set of
leaf chemical & structural parameters}
\usage{
prospect_lut(input_prospect, spec_prospect = NULL)
}
\arguments{
\item{input_prospect}{dataframe. list of PROSPECT input parameters.}

\item{spec_prospect}{list. spectral constants
refractive index, specific absorption coefficients & spectral bands}
}
\value{
list. LUT including leaf reflectance and transmittance
}
\description{
computation of a LUT of leaf optical properties using a set of
leaf chemical & structural parameters
}
