% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_lop_for_inversion.R
\name{reshape_lop_for_inversion}
\alias{reshape_lop_for_inversion}
\title{Function to reshape reflectance and transmittance, and convert into dataframes}
\usage{
reshape_lop_for_inversion(refl, tran, spec_prospect)
}
\arguments{
\item{refl}{numeric. measured reflectance data}

\item{tran}{numeric. measured transmittance data}

\item{spec_prospect}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}
}
\value{
RT list of leaf optics converted into dataframe and coresponding number of samples
}
\description{
Function to reshape reflectance and transmittance, and convert into dataframes
}
