\name{runShinySED}
\alias{runShinySED}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Runs Shiny SED App
}
\description{
Simple wrapper function to run the Shiny SED App.
}
\usage{
runShinySED()
}
%- maybe also 'usage' for other objects documented here.
\details{
Just run it! Should be obvious what it does.
}
\value{
Just makes a Shiny app run, which in turn creates an interactive SED plot.
}
\author{
Aaron Robotham
}

\seealso{
\code{\link{ProSpectSED}}
}
\examples{
\dontrun{
runShinySED()
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Shiny }% use one of  RShowDoc("KEYWORDS")
\keyword{ SED }% __ONLY ONE__ keyword per line
