% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_init_parm.R
\name{set_init_parm}
\alias{set_init_parm}
\title{Function to set initial parameterization when performing PROSPECT inversion
using optimal configuration}
\usage{
set_init_parm(
  parms_to_estimate,
  parm_estimated,
  prospect_version,
  Nprior,
  ant_init,
  optimal_domain,
  refl,
  tran
)
}
\arguments{
\item{parms_to_estimate}{character. Parameters to estimate}

\item{parm_estimated}{character. PROSPECT parameters to be estimated}

\item{prospect_version}{character. Version of prospect model used for the
inversion: 'D' or 'PRO'}

\item{Nprior}{numeric prior estimation of N}

\item{ant_init}{numeric prior estimation of ANT}

\item{optimal_domain}{vector. optimal spectral domain}

\item{refl}{dataframe reflectance data}

\item{tran}{dataframe transmittance data}
}
\value{
list containing parms_to_estimate, parm_estimated, ul_bounds, prospect_version.
parms_to_estimate_tmp, init_values
}
\description{
Function to set initial parameterization when performing PROSPECT inversion
using optimal configuration
}
