% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n_values.R
\name{set_n_values}
\alias{set_n_values}
\title{Function set N values if reflectance or transmittance is missing}
\usage{
set_n_values(refl, tran, spec_prospect)
}
\arguments{
\item{refl}{dataframe. reflectance data}

\item{tran}{dataframe. transmittance data}

\item{spec_prospect}{dataframe. spectral properties used in PROSPECT}
}
\value{
Nprior dataframe. N values corresponding to refl or tran
}
\description{
Function set N values if reflectance or transmittance is missing
}
