% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options_prospect.R
\name{set_options_prospect}
\alias{set_options_prospect}
\title{set options}
\usage{
set_options_prospect(fun, options = NULL)
}
\arguments{
\item{fun}{character. name of the function which has optional parameters}

\item{options}{list. including
\itemize{
\item nb_clusters numeric. number of clusters
}}
}
\value{
options with default values when missing
}
\description{
set options
}
