% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{SetInitParm}
\alias{SetInitParm}
\title{Function to set initial parameterization when performing PROSPECT inversion
using optimal configuration}
\usage{
SetInitParm(
  Parms2Estimate,
  ParmEst,
  PROSPECT_version,
  Nprior,
  ANTinit,
  OptDomain,
  Refl,
  Tran
)
}
\arguments{
\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{ParmEst}{character. PROSPECT parameters to be estimated}

\item{PROSPECT_version}{character. Version of prospect model used for the
inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB',}

\item{Nprior}{numeric prior estimation of N}

\item{ANTinit}{numeric prior estimation of ANT}

\item{OptDomain}{vector. optimal spectral domain}

\item{Refl}{dataframe reflectance data}

\item{Tran}{dataframe transmittance data}
}
\value{
list containing Parms2Estimate, ParmEst, UL_Bounds, PROSPECT_version.
Parms2Estimate_tmp, InitValues
}
\description{
Function to set initial parameterization when performing PROSPECT inversion
using optimal configuration
}
