% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{SetNValues}
\alias{SetNValues}
\title{Function set N values if reflectance or transmittance is missing}
\usage{
SetNValues(Refl, Tran, SpecPROSPECT)
}
\arguments{
\item{Refl}{dataframe. reflectance data}

\item{Tran}{dataframe. transmittance data}

\item{SpecPROSPECT}{dataframe. spectral properties used in PROSPECT}
}
\value{
Nprior dataframe. N values corresponding to Refl or Tran
}
\description{
Function set N values if reflectance or transmittance is missing
}
