% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_inversion.R
\name{try_inversion}
\alias{try_inversion}
\title{Function handling error during inversion}
\usage{
try_inversion(
  x0,
  merit_function,
  spec_prospect,
  refl,
  tran,
  parms_to_estimate,
  lb,
  ub,
  verbose = FALSE
)
}
\arguments{
\item{x0}{numeric. Vector of input variables to estimate}

\item{merit_function}{character. name of the merit function}

\item{spec_prospect}{list. Includes optical constants refractive index,
specific absorption coefficients and corresponding spectral bands}

\item{refl}{numeric. measured reflectance data}

\item{tran}{numeric. measured transmittance data}

\item{parms_to_estimate}{character vector. Parameters to estimate
to be estimated through inversion}

\item{lb}{numeric. Lower bound}

\item{ub}{numeric. Upper bound}

\item{verbose}{boolean. Set TRUE for information on adjustment of tolerance
during inversion.}
}
\value{
res estimates of the parameters
}
\description{
Function handling error during inversion
}
\details{
This function is based on \code{\link[pracma]{fmincon}}.
}
