% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{tryInversion}
\alias{tryInversion}
\title{Function handling error during inversion}
\usage{
tryInversion(
  x0,
  MeritFunction,
  SpecPROSPECT,
  Refl,
  Tran,
  Parms2Estimate,
  lb,
  ub
)
}
\arguments{
\item{MeritFunction}{character. name of the function to be used as merit function}

\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Refl}{numeric. measured reflectance data}

\item{Tran}{numeric. measured Transmittance data}

\item{lb}{numeric. Lower bound}

\item{ub}{numeric. Upper bound}

\item{xinit}{numeric. Vector of input variables to estimate}

\item{InPROSPECT}{dataframe. set of PROSPECT input variables}

\item{Vars2Estimate}{numeric. location of variables from Input_PROSPECT
to be estimated through inversion}
}
\value{
fc estimates of the parameters
}
\description{
Function handling error during inversion
}
