% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prospect_parms.R
\name{check_prospect_parms}
\alias{check_prospect_parms}
\title{Function to check good agreement between prospect version and parameter list}
\usage{
check_prospect_parms(
  prospect_version,
  parms_to_estimate,
  estimate_brown_pigments,
  estimate_alpha,
  xlub,
  init_values
)
}
\arguments{
\item{prospect_version}{character. version used for inversion: 'D' or 'PRO'}

\item{parms_to_estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{estimate_brown_pigments}{boolean. should brown pigments be accounted for during inversion?}

\item{estimate_alpha}{boolean. should alpha be accounted for during inversion?}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{parms_to_estimate} first
line being the lower boundaries and second line the upper boundaries.}

\item{init_values}{dataframe. Default values of PROSPECT parameters. During
optimization, they are used either as initialization values for parameters to
estimate, or as fix values for other parameters.
Parameters not compatible with prospect_version are not taken into account.}
}
\value{
list of parameters to estimate & corresponding lower/upper boundaries
}
\description{
Function to check good agreement between prospect version and parameter list
}
