% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{check_prospect_parms}
\alias{check_prospect_parms}
\title{Function to check good agreement between prospect version and parameter list}
\usage{
check_prospect_parms(
  PROSPECT_version,
  Parms2Estimate,
  Est_Brown_Pigments,
  Est_alpha,
  xlub,
  InitValues
)
}
\arguments{
\item{PROSPECT_version}{character. Version of prospect model used for the inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB',
See details.}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{Est_Brown_Pigments}{boolean. should brown pigments be accounted for during inversion?}

\item{Est_alpha}{boolean. should alpha be accounted for during inversion?}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate} first line being
the lower boundaries and second line the upper boundaries.}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}
}
\value{
list consolidated list of parameters to estimate and corresponding lower / upper boundaries
}
\description{
Function to check good agreement between prospect version and parameter list
}
