% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version_prospect.R
\name{check_version_prospect}
\alias{check_version_prospect}
\title{This function checks if the input parameters are defined as expected
to run either PROSPECT-D or PROSPECT-PRO}
\usage{
check_version_prospect(lma, prot, cbc)
}
\arguments{
\item{lma}{numeric. content corresponding to lma}

\item{prot}{numeric. content corresponding to protein content}

\item{cbc}{numeric. content corresponding to carbon based constituents}
}
\value{
list. updated lma, prot and cbc
}
\description{
This function checks if the input parameters are defined as expected
to run either PROSPECT-D or PROSPECT-PRO
}
