\name{filters}
\alias{filters}
\alias{filt_FUV_GALEX}
\alias{filt_NUV_GALEX}
\alias{filt_u_SDSS}
\alias{filt_u_VST}
\alias{filt_u_LSST}
\alias{filt_g_SDSS}
\alias{filt_g_VST}
\alias{filt_g_HSC}
\alias{filt_g_LSST}
\alias{filt_r_SDSS}
\alias{filt_r_VST}
\alias{filt_r_HSC}
\alias{filt_r_LSST}
\alias{filt_i_SDSS}
\alias{filt_i_VST}
\alias{filt_i_HSC}
\alias{filt_i_LSST}
\alias{filt_z_SDSS}
\alias{filt_Z_VISTA}
\alias{filt_Z_UKIRT}
\alias{filt_z_VST}
\alias{filt_z_HSC}
\alias{filt_z_LSST}
\alias{filt_Y_VISTA}
\alias{filt_Y_UKIRT}
\alias{filt_Y_HSC}
\alias{filt_y_LSST}
\alias{filt_J_VISTA}
\alias{filt_J_UKIRT}
\alias{filt_H_VISTA}
\alias{filt_H_UKIRT}
\alias{filt_K_VISTA}
\alias{filt_K_UKIRT}
\alias{filt_Ks_VISTA}
\alias{filt_W1_WISE}
\alias{filt_W2_WISE}
\alias{filt_W3_WISE}
\alias{filt_W4_WISE}
\alias{filt_I1_Spitzer}
\alias{filt_I2_Spitzer}
\alias{filt_I3_Spitzer}
\alias{filt_I4_Spitzer}
\alias{filt_M24_Spitzer}
\alias{filt_M70_Spitzer}
\alias{filt_M160_Spitzer}
\alias{filt_P70_Herschel}
\alias{filt_P100_Herschel}
\alias{filt_P160_Herschel}
\alias{filt_S250_Herschel}
\alias{filt_S350_Herschel}
\alias{filt_S500_Herschel}
\alias{filt_S450_JCMT}
\alias{filt_S850_JCMT}
\alias{filt_1mm_Aztec}
\alias{filt_2mm_Gizmo}
\alias{filt_F070W_JWST}
\alias{filt_F090W_JWST}
\alias{filt_F115W_JWST}
\alias{filt_F140M_JWST}
\alias{filt_F150W_JWST}
\alias{filt_F150W2_JWST}
\alias{filt_F162M_JWST}
\alias{filt_F164N_JWST}
\alias{filt_F182M_JWST}
\alias{filt_F187N_JWST}
\alias{filt_F200W_JWST}
\alias{filt_F210M_JWST}
\alias{filt_F212N_JWST}
\alias{filt_F250M_JWST}
\alias{filt_F277W_JWST}
\alias{filt_F300M_JWST}
\alias{filt_F322W2_JWST}
\alias{filt_F323N_JWST}
\alias{filt_F335M_JWST}
\alias{filt_F356W_JWST}
\alias{filt_F360M_JWST}
\alias{filt_F405N_JWST}
\alias{filt_F410M_JWST}
\alias{filt_F430M_JWST}
\alias{filt_F444W_JWST}
\alias{filt_F460M_JWST}
\alias{filt_F466N_JWST}
\alias{filt_F470N_JWST}
\alias{filt_F480M_JWST}
\alias{filt_F560W_JWST}
\alias{filt_F770W_JWST}
\alias{filt_F1000W_JWST}
\alias{filt_F1130W_JWST}
\alias{filt_F1280W_JWST}
\alias{filt_F1500W_JWST}
\alias{filt_F1800W_JWST}
\alias{filt_F2100W_JWST}
\alias{filt_F2550W_JWST}
\alias{filt_Blue_Euclid}
\alias{filt_H_Euclid}
\alias{filt_J_Euclid}
\alias{filt_Red_Euclid}
\alias{filt_VIS_Euclid}
\alias{filt_Y_Euclid}
\alias{filt_F184_WFIRST}
\alias{filt_H158_WFIRST}
\alias{filt_J129_WFIRST}
\alias{filt_R062_WFIRST}
\alias{filt_W146_WFIRST}
\alias{filt_Y106_WFIRST}
\alias{filt_Z087_WFIRST}
\alias{filt_Band9_ALMA}
\alias{filt_Band8_ALMA}
\alias{filt_Band7_ALMA}
\alias{filt_Band6_ALMA}
\alias{filt_Band5_ALMA}
\alias{filt_Band4_ALMA}
\alias{filt_Band3_ALMA}
\alias{filt_Band2_ALMA}
\alias{filt_Band1_ALMA}
\alias{filt_BandQ_VLA}
\alias{filt_BandKa_VLA}
\alias{filt_BandK_VLA}
\alias{filt_BandKu_VLA}
\alias{filt_BandX_VLA}
\alias{filt_BandC_VLA}
\alias{filt_BandS_VLA}
\alias{filt_BandL_VLA}
\alias{filt_BandP_VLA}
\alias{filt_Band5_GMRT}
\alias{filt_Band4_GMRT}
\alias{filt_Band3_GMRT}
\alias{filt_Band2_GMRT}
\alias{filt_F062_ROMAN}
\alias{filt_F087_ROMAN}
\alias{filt_F106_ROMAN}
\alias{filt_F129_ROMAN}
\alias{filt_F146_ROMAN}
\alias{filt_F158_ROMAN}
\alias{filt_F184_ROMAN}
\alias{filt_F213_ROMAN}
\alias{filt_W146_ROMAN}
\alias{filt_COBE_DIRBE_60m}
\alias{filt_COBE_DIRBE_100m}
\alias{filt_COBE_DIRBE_140m}
\alias{filt_COBE_DIRBE_240m}
\alias{filt_PLANCK_HFI_100GHz}
\alias{filt_PLANCK_HFI_143GHz}
\alias{filt_PLANCK_HFI_217GHz}
\alias{filt_PLANCK_HFI_353GHz}
\alias{filt_PLANCK_HFI_545GHz}
\alias{filt_PLANCK_HFI_857GHz}
\alias{filt}
\alias{filter}
\alias{filters}
\docType{data}
\title{
Band Filters
}
\description{
Various common filter response functions, appropriate for using in calculating magitudes etc.
}
\usage{
data("filt_FUV_GALEX")
data("filt_NUV_GALEX")
data("filt_u_SDSS")
data("filt_g_SDSS")
data("filt_r_SDSS")
data("filt_i_SDSS")
data("filt_z_SDSS")
data("filt_u_VST")
data("filt_g_VST")
data("filt_r_VST")
data("filt_i_VST")
data("filt_z_VST")
data("filt_u_LSST")
data("filt_g_LSST")
data("filt_r_LSST")
data("filt_i_LSST")
data("filt_z_LSST")
data("filt_y_LSST")
data("filt_g_HSC")
data("filt_r_HSC")
data("filt_i_HSC")
data("filt_z_HSC")
data("filt_Y_HSC")
data("filt_Z_VISTA")
data("filt_Y_VISTA")
data("filt_J_VISTA")
data("filt_H_VISTA")
data("filt_K_VISTA")
data("filt_Ks_VISTA")
data("filt_Z_UKIRT")
data("filt_Y_UKIRT")
data("filt_J_UKIRT")
data("filt_H_UKIRT")
data("filt_K_UKIRT")
data("filt_W1_WISE")
data("filt_W2_WISE")
data("filt_W3_WISE")
data("filt_W4_WISE")
data("filt_I1_Spitzer")
data("filt_I2_Spitzer")
data("filt_I3_Spitzer")
data("filt_I4_Spitzer")
data("filt_M24_Spitzer")
data("filt_M70_Spitzer")
data("filt_M160_Spitzer")
data("filt_P70_Herschel")
data("filt_P100_Herschel")
data("filt_P160_Herschel")
data("filt_S250_Herschel")
data("filt_S350_Herschel")
data("filt_S500_Herschel")
data("filt_S450_JCMT")
data("filt_S850_JCMT")
data("filt_1mm_Aztec")
data("filt_2mm_Gizmo")
data("filt_F070W_JWST")
data("filt_F090W_JWST")
data("filt_F115W_JWST")
data("filt_F140M_JWST")
data("filt_F150W_JWST")
data("filt_F150W2_JWST")
data("filt_F162M_JWST")
data("filt_F164N_JWST")
data("filt_F182M_JWST")
data("filt_F187N_JWST")
data("filt_F200W_JWST")
data("filt_F210M_JWST")
data("filt_F212N_JWST")
data("filt_F250M_JWST")
data("filt_F277W_JWST")
data("filt_F300M_JWST")
data("filt_F322W2_JWST")
data("filt_F323N_JWST")
data("filt_F335M_JWST")
data("filt_F356W_JWST")
data("filt_F360M_JWST")
data("filt_F405N_JWST")
data("filt_F410M_JWST")
data("filt_F430M_JWST")
data("filt_F444W_JWST")
data("filt_F460M_JWST")
data("filt_F466N_JWST")
data("filt_F470N_JWST")
data("filt_F480M_JWST")
data("filt_F560W_JWST")
data("filt_F770W_JWST")
data("filt_F1000W_JWST")
data("filt_F1130W_JWST")
data("filt_F1280W_JWST")
data("filt_F1500W_JWST")
data("filt_F1800W_JWST")
data("filt_F2100W_JWST")
data("filt_F2550W_JWST")
data("filt_Blue_Euclid")
data("filt_H_Euclid")
data("filt_J_Euclid")
data("filt_Red_Euclid")
data("filt_VIS_Euclid")
data("filt_Y_Euclid")
data("filt_F184_WFIRST")
data("filt_H158_WFIRST")
data("filt_J129_WFIRST")
data("filt_R062_WFIRST")
data("filt_W146_WFIRST")
data("filt_Y106_WFIRST")
data("filt_Z087_WFIRST")
data("filt_Band9_ALMA")
data("filt_Band8_ALMA")
data("filt_Band7_ALMA")
data("filt_Band6_ALMA")
data("filt_Band5_ALMA")
data("filt_Band4_ALMA")
data("filt_Band3_ALMA")
data("filt_Band2_ALMA")
data("filt_Band1_ALMA")
data("filt_BandQ_VLA")
data("filt_BandKa_VLA")
data("filt_BandK_VLA")
data("filt_BandKu_VLA")
data("filt_BandX_VLA")
data("filt_BandC_VLA")
data("filt_BandS_VLA")
data("filt_BandL_VLA")
data("filt_BandP_VLA")
data("filt_Band5_GMRT")
data("filt_Band4_GMRT")
data("filt_Band3_GMRT")
data("filt_Band2_GMRT")
data("filt_F062_ROMAN")
data("filt_F087_ROMAN")
data("filt_F106_ROMAN")
data("filt_F129_ROMAN")
data("filt_F146_ROMAN")
data("filt_F158_ROMAN")
data("filt_F184_ROMAN")
data("filt_F213_ROMAN")
data("filt_W146_ROMAN")
data("filt_COBE_DIRBE_60m")
data("filt_COBE_DIRBE_100m")
data("filt_COBE_DIRBE_140m")
data("filt_COBE_DIRBE_240m")
data("filt_PLANCK_HFI_100GHz")
data("filt_PLANCK_HFI_143GHz")
data("filt_PLANCK_HFI_217GHz")
data("filt_PLANCK_HFI_353GHz")
data("filt_PLANCK_HFI_545GHz")
data("filt_PLANCK_HFI_857GHz")
}
\format{
  A number of data frames with the following 2 variables.
  \describe{
    \item{\code{wave}}{A numeric vector, wavelength in Angstroms}
    \item{\code{response}}{A numeric vector, the relative response}
  }
}
\details{
Note the filter for VISTA is actually Ks (for K-short) which makes it approximately similar to 2MASS. This differs to WFCAM (e.g. UKIDSS) which used a full K filter. Both K_VISTA and Ks_VISTA are provided to prevent silly name convention mistakes, but these response curves are in fact identical (i.e. Ks).

You can get additional (obscure) filters at \url{http://svo2.cab.inta-csic.es/theory/fps/index.php}.
}
\source{
Mostly the relevant websites for various facilities, but in particular:

\url{http://svo2.cab.inta-csic.es/theory/fps/index.php}
}
\examples{
colvec=rev(rainbow(9, end=2/3))
plot(filt_u_VST, type='l', col=colvec[1], xlim=c(3e3,2.5e4), ylim=c(0,1),
xlab = 'Wave / Ang', ylab = 'Response')
lines(filt_g_VST, type='l', col=colvec[2])
lines(filt_r_VST, type='l', col=colvec[3])
lines(filt_i_VST, type='l', col=colvec[4])
lines(filt_Z_VISTA, type='l', col=colvec[5])
lines(filt_Y_VISTA, type='l', col=colvec[6])
lines(filt_J_VISTA, type='l', col=colvec[7])
lines(filt_H_VISTA, type='l', col=colvec[8])
lines(filt_K_VISTA, type='l', col=colvec[9])
}
\keyword{datasets}
