% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Invert_PROSPECT}
\alias{Invert_PROSPECT}
\title{Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
Invert_PROSPECT(
  SpecPROSPECT,
  Refl = NULL,
  Tran = NULL,
  InitValues = data.frame(CHL = 40, CAR = 10, ANT = 0.1, BROWN = 0.01, EWT = 0.01, LMA
    = 0.01, PROT = 0.001, CBC = 0.009, N = 1.5, alpha = 40),
  Parms2Estimate = "ALL",
  PROSPECT_version = "D",
  MeritFunction = "Merit_RMSE_PROSPECT",
  xlub = data.frame(CHL = c(1e-04, 150), CAR = c(1e-04, 25), ANT = c(0, 50), BROWN =
    c(0, 1), EWT = c(1e-07, 0.08), LMA = c(1e-06, 0.04), PROT = c(1e-07, 0.005), CBC =
    c(1e-06, 0.04), N = c(0.5, 4), alpha = c(10, 90)),
  alphaEst = FALSE
)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{PROSPECT_version}{character. Version of prospect model used for the inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB',
See details.}

\item{MeritFunction}{character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate} first line being
the lower boundaries and second line the upper boundaries.}

\item{alphaEst}{boolean. should alpha be estimated or not?}
}
\value{
OutPROSPECT estimated values corresponding to Parms2Estimate
}
\description{
Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
\details{
Six versions of prospect are available for inversion.
The version is depending on the parameters taken into account:\tabular{ccccccc}{
   Version \tab 5 \tab 5B \tab D \tab DB \tab PRO \tab PROB \cr
   CHL \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \cr
   CAR \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \cr
   ANT \tab  \tab  \tab ✅ \tab ✅ \tab ✅ \tab ✅ \cr
   BROWN \tab  \tab ✅ \tab  \tab ✅ \tab  \tab ✅ \cr
   EWT \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \cr
   LMA \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab  \tab  \cr
   PROT \tab  \tab  \tab  \tab  \tab ✅ \tab ✅ \cr
   CBC \tab  \tab  \tab  \tab  \tab ✅ \tab ✅ \cr
   N \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \tab ✅ \cr
}


Argument \code{InitValues} is expecting a default value for each of the parameters as well as an \code{alpha} value.
}
