% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Invert_PROSPECT_OPT}
\alias{Invert_PROSPECT_OPT}
\title{This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only Reflectance or only Transmittance is available}
\usage{
Invert_PROSPECT_OPT(
  SpecPROSPECT,
  lambda,
  Refl = NULL,
  Tran = NULL,
  PROSPECT_version = "D",
  Parms2Estimate = "ALL",
  InitValues = data.frame(CHL = 40, CAR = 10, ANT = 0.1, BROWN = 0.01, EWT = 0.01, LMA
    = 0.01, PROT = 0.001, CBC = 0.009, N = 1.5, alpha = 40)
)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{lambda}{numeric. spectral bands corresponding to experimental data}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}

\item{PROSPECT_version}{character. Version of prospect model used for the inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB',
See details.}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}
}
\value{
Nprior vector corresponding to teh prior estimation of N based on R only or T only
}
\description{
This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only Reflectance or only Transmittance is available
}
