% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Invert_PROSPECT_OPT}
\alias{Invert_PROSPECT_OPT}
\title{This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only Reflectance or only Transmittance is available}
\usage{
Invert_PROSPECT_OPT(
  lambda,
  SpecPROSPECT = NULL,
  Refl = NULL,
  Tran = NULL,
  PROSPECT_version = "D",
  Parms2Estimate = "ALL",
  InitValues = data.frame(CHL = 40, CAR = 10, ANT = 0.1, BROWN = 0.01, EWT = 0.01, LMA =
    0.01, PROT = 0.001, CBC = 0.009, N = 1.5, alpha = 40),
  xlub = data.frame(CHL = c(1e-04, 150), CAR = c(1e-04, 25), ANT = c(0, 50), BROWN = c(0,
    1), EWT = c(1e-08, 0.1), LMA = c(1e-06, 0.06), PROT = c(1e-07, 0.006), CBC = c(1e-06,
    0.054), N = c(0.5, 4), alpha = c(10, 90)),
  verbose = FALSE,
  progressBar = TRUE
)
}
\arguments{
\item{lambda}{numeric. spectral bands corresponding to experimental data}

\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}

\item{PROSPECT_version}{character. Version of prospect model used for the
inversion: 'D', 'PRO'.}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate} first line being
the lower boundaries and second line the upper boundaries.}

\item{verbose}{boolean. set to TRUE to display sample number to be inverted
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}

\item{progressBar}{boolean. show progressbar?}
}
\value{
Nprior vector corresponding to teh prior estimation of N based on R only or T only
}
\description{
This function uses optimal configuration identified by Spafford et al. (2020) to estimate leaf chemistry
prior information on N is provided if only Reflectance or only Transmittance is available
}
