% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Invert_PROSPECT_subdomain}
\alias{Invert_PROSPECT_subdomain}
\title{Function performing PROSPECT inversion on a spectral subset combining an
initial feature subset (Initial_Features) and an additional feature subset (New_Features)}
\usage{
Invert_PROSPECT_subdomain(
  New_Features,
  Refl,
  Tran,
  SpecPROSPECT,
  lambda,
  BiochTruth,
  Target,
  Parms2Estimate,
  Initial_Features,
  InitValues = data.frame(CHL = 40, CAR = 10, ANT = 0.1, BROWN = 0.01, EWT = 0.01, LMA
    = 0.01, PROT = 0.001, CBC = 0.009, N = 1.5, alpha = 40),
  PROSPECT_version = "D",
  MeritFunction = "Merit_RMSE_PROSPECT",
  alphaEst = FALSE,
  xlub = data.frame(CHL = c(1e-04, 150), CAR = c(1e-04, 25), ANT = c(0, 50), BROWN =
    c(0, 1), EWT = c(1e-08, 0.1), LMA = c(1e-06, 0.06), PROT = c(1e-07, 0.006), CBC =
    c(1e-06, 0.054), N = c(0.5, 4), alpha = c(10, 90)),
  verbose = FALSE
)
}
\arguments{
\item{New_Features}{spectral bands (in nm) to be added to the Initial_Features}

\item{Refl}{numeric. matrix of reflectances (n spectral bands x p samples)}

\item{Tran}{numeric. matrix of transmittances (n spectral bands x p samples)}

\item{SpecPROSPECT}{list. Optical constants for a spectral domain broader than spectral_domain}

\item{lambda}{numeric. spectral bands corresponding to reflectance and transmittance measurements}

\item{BiochTruth}{numeric. value of biophysical/biochemical parameter to estimate for each sample}

\item{Target}{character. name of the parameter.
Should be picked between "CHL", "CAR", "ANT", "BROWN", "EWT", "PROT", "CBC", "N"}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{Initial_Features}{Initial feature set (in nm)}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}

\item{PROSPECT_version}{character. Version of prospect model used for the inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB'}

\item{MeritFunction}{character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)}

\item{alphaEst}{boolean. should alpha be estimated or not?}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate} first line being
the lower boundaries and second line the upper boundaries.}

\item{verbose}{boolean. set to TRUE to display sample number to be inverted}
}
\value{
list containing estimated value of the target and RMSE compared to measured value
}
\description{
Function performing PROSPECT inversion on a spectral subset combining an
initial feature subset (Initial_Features) and an additional feature subset (New_Features)
}
