% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{optimal_features_SFS}
\alias{optimal_features_SFS}
\title{Function performing optimal feature selection based on sequential forward feature slection}
\usage{
optimal_features_SFS(
  Refl = NULL,
  Tran = NULL,
  lambda,
  BiochTruth,
  Target,
  Parms2Estimate = "ALL",
  InitValues = data.frame(CHL = 40, CAR = 10, ANT = 0.1, BROWN = 0.01, EWT = 0.01, LMA =
    0.01, PROT = 0.001, CBC = 0.009, N = 1.5, alpha = 40),
  xlub = data.frame(CHL = c(1e-04, 150), CAR = c(1e-04, 25), ANT = c(0, 50), BROWN = c(0,
    1), EWT = c(1e-08, 0.1), LMA = c(1e-06, 0.06), PROT = c(1e-07, 0.006), CBC = c(1e-06,
    0.054), N = c(0.5, 4), alpha = c(10, 90)),
  SpecPROSPECT,
  spectral_domain,
  spectral_width,
  number_features,
  PROSPECT_version = "D",
  MeritFunction = "Merit_PROSPECT_RMSE",
  Est_alpha = FALSE,
  verbose = FALSE,
  nbCPU = 1,
  Continue = FALSE,
  AlreadyDone = NULL
)
}
\arguments{
\item{Refl}{numeric. matrix of reflectances (n spectral bands x p samples)}

\item{Tran}{numeric. matrix of transmittances (n spectral bands x p samples)}

\item{lambda}{numeric. spectral bands corresponding to reflectance and transmittance measurements}

\item{BiochTruth}{numeric. value of biophysical/biochemical parameter to estimate for each sample}

\item{Target}{character. name of the parameter.
Should be picked between "CHL", "CAR", "ANT", "BROWN", "EWT", "PROT", "CBC", "N"}

\item{Parms2Estimate}{character vector. Parameters to estimate (can be 'ALL')}

\item{InitValues}{data.frame. Default values of PROSPECT parameters. During optimization,
they are used either as initialization values for parameters to estimate,
or as fix values for other parameters.
Parameters not compatible with PROSPECT_version are not taken into account.}

\item{xlub}{data.frame. Boundaries of the parameters to estimate.
The data.frame must have columns corresponding to \code{Parms2Estimate} first line being
the lower boundaries and second line the upper boundaries.}

\item{SpecPROSPECT}{list. Optical constants for a spectral domain broader than spectral_domain}

\item{spectral_domain}{vector. defines minimum and maximum wavelengths of the spectral domain (in nm).
Assumes 1nm spectral sampling}

\item{spectral_width}{vector. width of individual spectral features (in nm)}

\item{number_features}{vector. number of features to be identified}

\item{PROSPECT_version}{character. Version of prospect model used for the inversion: '5', '5B', 'D', 'DB', 'PRO', 'PROB',}

\item{MeritFunction}{character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)}

\item{Est_alpha}{boolean. should alpha be estimated or not?}

\item{verbose}{boolean. set to TRUE to display sample number to be inverted}

\item{nbCPU}{numeric. defines number of CPU for multithread processing}

\item{Continue}{boolean. set to TRUE if the function has already been run for a lower value of number_features}

\item{AlreadyDone}{list. contains output of function optimal_features_SFS previouly run with lower value of number_features}
}
\value{
list containing :
\itemize{
\item SpectralFeatures = the Spectral Features identified as most relevant for estimation of Target (vector)
\item SpectralFeatures_List = the Spectral Features identified as most relevant for estimation of Target (list)
\item EstimatedParm = the estimated Target parameter for each additional spectral feature
\item RMSE = the minimum RMSE corresponding to the estimation of Target for each asditional spectral feature
}
}
\description{
Function performing optimal feature selection based on sequential forward feature slection
}
