% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect-package.R
\docType{package}
\name{prospect-package}
\alias{prospect}
\alias{prospect-package}
\title{prospect: PROSPECT leaf radiative transfer model and inversion routines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package prospect includes PROSPECT-PRO, the latest version of the PROSPECT leaf model. It also includes inversion routines based on iterative optimization in order to estimate leaf chemistry from spectroscopic measurements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/jbferet/prospect}
  \item Report bugs at \url{https://gitlab.com/jbferet/prospect/issues}
}

}
\author{
\strong{Maintainer}: Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (\href{https://orcid.org/0000-0002-0151-1334}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Florian de Boissieu \email{florian.deboissieu@irstea.fr} (\href{https://orcid.org/0000-0002-2185-9952}{ORCID})
}

}
\keyword{internal}
