% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{prospect}
\alias{prospect}
\title{core function running PROSPECT
This function allows simulations using PROSPECT-D or PROSPECT-PRO depending
on the parameterization.}
\usage{
prospect(
  spec_prospect = NULL,
  input_prospect = NULL,
  n_struct = 1.5,
  chl = 40,
  car = 8,
  ant = 0,
  brown = 0,
  ewt = 0.01,
  lma = NULL,
  prot = 0,
  cbc = 0,
  alpha = 40,
  check = TRUE
)
}
\arguments{
\item{spec_prospect}{list. Includes spectral constants derived from
spec_prospect_full_range: refractive index, specific absorption coefficients
and corresponding spectral bands}

\item{input_prospect}{list. Includes all prospect input parameters}

\item{n_struct}{numeric. Leaf structure parameter}

\item{chl}{numeric. Chlorophyll content (microg.cm-2)}

\item{car}{numeric. Carotenoid content (microg.cm-2)}

\item{ant}{numeric. Anthocyanin content (microg.cm-2)}

\item{brown}{numeric. Brown pigment content (Arbitrary units)}

\item{ewt}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{lma}{numeric. Leaf Mass per Area (g.cm-2)}

\item{prot}{numeric. protein content  (g.cm-2)}

\item{cbc}{numeric. NonProt Carbon-based constituent content (g.cm-2)}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf}

\item{check}{boolean. set to TRUE to check input data format}
}
\value{
leaf directional-hemispherical reflectance and transmittance
}
\description{
core function running PROSPECT
This function allows simulations using PROSPECT-D or PROSPECT-PRO depending
on the parameterization.
}
