% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{PROSPECT_LUT}
\alias{PROSPECT_LUT}
\title{computation of a LUT of leaf optical properties using a set of
leaf chemical & structural parameters}
\usage{
PROSPECT_LUT(Input_PROSPECT, SpecPROSPECT = NULL)
}
\arguments{
\item{Input_PROSPECT}{dataframe. full list of PROSPECT input parameters. Default = Set to NULL}

\item{SpecPROSPECT}{list. spectral constants
refractive index, specific absorption coefficients & spectral bands}
}
\value{
list. LUT including leaf reflectance and transmittance
}
\description{
computation of a LUT of leaf optical properties using a set of
leaf chemical & structural parameters
}
