% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{tryInversion}
\alias{tryInversion}
\title{Function handling error during inversion}
\usage{
tryInversion(
  x0,
  MeritFunction,
  SpecPROSPECT,
  Refl,
  Tran,
  Parms2Estimate,
  lb,
  ub,
  verbose = FALSE
)
}
\arguments{
\item{x0}{numeric. Vector of input variables to estimate}

\item{MeritFunction}{character. name of the merit function}

\item{SpecPROSPECT}{list. Includes optical constants refractive index,
specific absorption coefficients and corresponding spectral bands}

\item{Refl}{numeric. measured reflectance data}

\item{Tran}{numeric. measured Transmittance data}

\item{Parms2Estimate}{character vector. Parameters to estimate
to be estimated through inversion}

\item{lb}{numeric. Lower bound}

\item{ub}{numeric. Upper bound}

\item{verbose}{boolean. Set TRUE for information on adjustment of tolerance
during inversion.}
}
\value{
res estimates of the parameters
}
\description{
Function handling error during inversion
}
\details{
This function is based on \code{\link[pracma]{fmincon}}.
}
